Merge All work

+++ Life/Source/sys_dbm.cpp
@@ -1,8 +1,8 @@
 /* Copyright by Denys Duchier, Jul 1995
    Simon Fraser University
 
-  DBM INTERFACE
-  */
+   DBM INTERFACE
+*/
 /*	$Id: sys_dbm.c,v 1.4 1996/01/17 00:34:34 duchier Exp $	*/
 
 #include "extern.h"
@@ -11,9 +11,9 @@
 #include "sys.h"
 #include <dbm.h>
 
-static long
+static long long
 dbminit_internal(args,result,funct)
-     ptr_psi_term args[],result,funct;
+  ptr_psi_term args[],result,funct;
 {
   if (dbminit((char*)args[0]->value)<0) return FALSE;
   else {
@@ -22,7 +22,7 @@ dbminit_internal(args,result,funct)
   }
 }
 
-static long
+static long long
 c_dbminit()
 {
   psi_arg args[1];
@@ -30,9 +30,9 @@ c_dbminit()
   return call_primitive(dbminit_internal,NARGS(args),args,0);
 }
 
-static long
+static long long
 dbmfetch_internal(args,result,funct)
-     ptr_psi_term args[],result,funct;
+  ptr_psi_term args[],result,funct;
 {
   datum d;
   d.dptr  = (char*)args[0]->value;
@@ -46,7 +46,7 @@ dbmfetch_internal(args,result,funct)
   }
 }
 
-static long
+static long long
 c_dbmfetch()
 {
   psi_arg args[1];
@@ -54,9 +54,9 @@ c_dbmfetch()
   return call_primitive(dbmfetch_internal,NARGS(args),args,0);
 }
 
-static long
+static long long
 dbmstore_internal(args,result,funct)
-     ptr_psi_term args[],result,funct;
+  ptr_psi_term args[],result,funct;
 {
   datum key,content;
   key.dptr  = (char*)args[0]->value;
@@ -67,7 +67,7 @@ dbmstore_internal(args,result,funct)
   else return TRUE;
 }
 
-static long
+static long long
 c_dbmstore()
 {
   psi_arg args[2];
@@ -76,9 +76,9 @@ c_dbmstore()
   return call_primitive(dbmstore_internal,NARGS(args),args,0);
 }
 
-static long
+static long long
 dbmdelete_internal(args,result,funct)
-     ptr_psi_term args[],result,funct;
+  ptr_psi_term args[],result,funct;
 {
   datum key;
   key.dptr  = (char*)args[0]->value;
@@ -87,7 +87,7 @@ dbmdelete_internal(args,result,funct)
   else return TRUE;
 }
 
-static long
+static long long
 c_dbmdelete()
 {
   psi_arg args[1];
@@ -95,9 +95,9 @@ c_dbmdelete()
   return call_primitive(dbmdelete_internal,NARGS(args),args,0);
 }
 
-static long
+static long long
 dbmfirstkey_internal(args,result,funct)
-     ptr_psi_term args[],result,funct;
+  ptr_psi_term args[],result,funct;
 {
   datum key;
   key=firstkey();
@@ -109,15 +109,15 @@ dbmfirstkey_internal(args,result,funct)
   }
 }
 
-static long
+static long long
 c_dbmfirstkey()
 {
   return call_primitive(dbmfirstkey_internal,0,NULL,0);
 }
 
-static long
+static long long
 dbmnextkey_internal(args,result,funct)
-     ptr_psi_term args[],result,funct;
+  ptr_psi_term args[],result,funct;
 {
   datum key;
   key.dptr  = (char*)args[0]->value;
@@ -131,7 +131,7 @@ dbmnextkey_internal(args,result,funct)
   }
 }
 
-static long
+static long long
 c_dbmnextkey()
 {
   psi_arg args[1];
